Refactor: Refactor directory structure
Create Course directory
