+++ frontend/src/Course/Header.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { ICourse } from "../index";
+import { ICourse } from "../../index";
 
 const Header: React.FC<{ courseData: ICourse }> = ({ courseData }) => {
     return <h1>{courseData.name}</h1>;
